packages/utils/prom-check-video-stream/requirements.txt
prometheus_async
aiohttp
ffmpeg-python
ipython
asyncio
